Added new game page and game page
