+++ Source/LudoAPI/Data/BoardRepo.cs
@@ -89,13 +89,4 @@ namespace Ludo.API.Data
             return _context.Board.Any(e => e.Id == id);
         }
     }
-
-    public interface IBoardRepo
-    {
-        Task<List<Board>> GetAllBoards();
-        Task<Board> GetBoardByName(string name);
-        Task<Task> UpdateBoard(int id, Board board);
-        Task<Task> AddBoard(Board board);
-        Task<Task> DeleteBoard(string name);
-    }
 }
