+++ Source/UI/Pages/Game.cshtml
@@ -1,0 +1,7 @@
+﻿@page "{boardname?}"
+@model Ludo.Web.Pages.GameModel
+@{
+}
+
+<h2>@Model.BoardName</h2>
+<h3>Players: @Model.Players</h3>
